Use message's receivedAt for sorting, clustering and displaying timestamps on conversations and messages.
Atlas used last message's sentAt for sorting conversations, clustering messages and showing timestamps. But sentAt is not set on a message until it is synced to server. So, this is absent in UI when offline, conversations are not in correct sorted order, and messages are not clustered properly.

With 0.20.5, layer sdk sets receivedAt to reflect the time the message was received by this user. So, when an user sends a message, it will have the current device timestamp on it. On receivers side, it will have the timestamp the message reached the server (and hence the user's inbox). This change allows UI to use receivedAt for sorting, clustering and displaying timestamp's everywhere and makes the user experience better.